+++ HISTORY.md
@@ -1,1 +1,7 @@
 # Changelog
+
+- Added api-token support
+- Added search route
+- Added index route
+- Added config file configuration
+- Added CLI